requirements.txt
pandas
numpy
plotly
openai
streamlit
python-dotenv
fastapi
uvicorn
scipy